+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Updated `isMobilePhone()` to match any locale's pattern by default
   ([#874](https://github.com/chriso/validator.js/pull/874))
+- Added an option to ignore whitespace in `isEmpty()`
+  ([#880](https://github.com/chriso/validator.js/pull/880))
 - New and improved locales
   ([#878](https://github.com/chriso/validator.js/pull/878),
    [#879](https://github.com/chriso/validator.js/pull/879))